refactory(HomePage): rename component
- 컴포넌트의 사용 용도에 더 맞는 이름으로 변경(HomePage -> CategoryList)
